Move Text API and implementation to Text world object
