+++ pygfx/geometries/__init__.py
@@ -42,15 +42,6 @@ The standardized names are:
     plane_geometry
     Geometry
 
-.. rubric:: Text Geometry
-.. autosummary::
-    :toctree: geometry/
-    :template: ../_templates/custom_layout.rst
-
-    TextGeometry
-    MultiTextGeometry
-    TextBlock
-
 .. rubric:: Other/Special Geometry
 .. autosummary::
     :toctree: geometry/
@@ -81,7 +72,6 @@ from ._polyhedron import (
     tetrahedron_geometry,
 )
 from ._toroidal import torus_knot_geometry, klein_bottle_geometry
-from ._text import TextGeometry, MultiTextGeometry, TextBlock
 
 # Define __all__ for e.g. Sphinx
 __all__ = [
@@ -93,5 +83,4 @@ __all__ = [
 __all__.sort()
 __all__.remove("Geometry")
 __all__.insert(0, "Geometry")
-__all__.insert(__all__.index("MultiTextGeometry") + 1, "TextBlock")
 __all__.append("geometry_from_trimesh")
