+++ pygfx/objects/__init__.py
@@ -18,6 +18,8 @@
     Image
     Volume
     Text
+    MultiText
+    TextBlock
     Ruler
     InstancedMesh
     Light
@@ -74,7 +76,7 @@ from ._more import (
     Volume,
 )
 from ._ruler import Ruler
-from ._text import Text
+from ._text import Text, MultiText, TextBlock
 from ._instanced import InstancedMesh
 from ._lights import Light, PointLight, DirectionalLight, AmbientLight, SpotLight
 from ._lights import (
